完成问题提问页面

+++ src/main/java/life/ma/jiang/community/dto/QuestionDTO.java
@@ -5,15 +5,15 @@ import lombok.Data;
 
 @Data
 public class QuestionDTO {
-    private int id;
+    private Long id;
     private String title;
     private  String description;
     private  Long gmtCreate;
     private Long gmtModified;
     private String tag;
-    private  int creator;
-    private int viewCount;
-    private int commentCount;
-    private int likeCount;
+    private  Long creator;
+    private Long viewCount;
+    private Long commentCount;
+    private Long likeCount;
     private User user;
 }
